refactor(profile): move the reorder drag out of the grid
The grid was doing three jobs: packing tiles, running a pointer gesture, and
reordering from the keyboard. The gesture is the biggest and the least to do
with the other two — a floating copy under the pointer, the previewed order,
the autoscroll, and the geometry that turns a point into an index.

It moves to useProfileBentoDrag, which takes the cards, the grid element and
its width, and calls back on drop. The grid keeps the packer and the keyboard
moves, and drops from 686 lines to 383.

A drag in flight when the component unmounts now takes its window listeners
and the grabbing cursor with it, which the inline version leaked.
